Move cluster metrics to other metrics
